requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
uvloop==0.19.0
alembic==1.13.1
PyYAML==6.0.1
aiohttp==3.9.1
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.11.1
pytest-timeout==2.1.0
httpx==0.25.2
hypothesis==6.82.1

# Code quality
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1

# Test containers (optional)
testcontainers==3.7.1

